requirements.txt
streamlit>=1.38.0
google-generativeai>=0.8.0
python-dotenv>=1.0.1
altair>=5.3.0
plotly>=5.24.0
